langchain/requirements.txt
langchain
langchain-community
langchain-dashscope
dashscope
langgraph
python-dotenv
chromadb

